Make available latest metadata for Tomte
