+++ cg/meta/workflow/rnafusion.py
@@ -8,7 +8,6 @@ from cg.constants.constants import Strandedness
 from cg.constants.nf_analysis import MULTIQC_NEXFLOW_CONFIG, RNAFUSION_METRIC_CONDITIONS
 from cg.exc import MissingMetrics
 from cg.meta.workflow.nf_analysis import NfAnalysisAPI
-from cg.models.analysis import NextflowAnalysis
 from cg.models.cg_config import CGConfig
 from cg.models.deliverables.metric_deliverables import MetricsBase
 from cg.models.rnafusion.rnafusion import RnafusionParameters, RnafusionSampleSheetEntry
@@ -115,21 +114,6 @@ class RnafusionAnalysisAPI(NfAnalysisAPI):
         if missing_metrics:
             LOG.error(f"Some mandatory metrics are missing: {', '.join(missing_metrics)}")
             raise MissingMetrics()
-
-    def parse_analysis(self, qc_metrics_raw: list[MetricsBase], **kwargs) -> NextflowAnalysis:
-        """Parse Rnafusion output analysis files and return analysis model."""
-        sample_metrics: dict[str, dict] = {}
-        for metric in qc_metrics_raw:
-            try:
-                sample_metrics[metric.id].update({metric.name.lower(): metric.value})
-            except KeyError:
-                sample_metrics[metric.id] = {metric.name.lower(): metric.value}
-        return NextflowAnalysis(sample_metrics=sample_metrics)
-
-    def get_latest_metadata(self, case_id: str) -> NextflowAnalysis:
-        """Return the latest metadata of a specific Rnafusion case."""
-        qc_metrics: list[MetricsBase] = self.get_multiqc_json_metrics(case_id)
-        return self.parse_analysis(qc_metrics_raw=qc_metrics)
 
     def get_workflow_metrics(self) -> dict:
         return RNAFUSION_METRIC_CONDITIONS
